refactor: split nine app lib hosts over 251 lines for wave-21 Slice B
Concern-split research breakdown, diplomacy row, last-turn playback, turn-feed locate, naval combine, tile radial chrome, counsel callbacks, work-order tooltips, and market tab chrome so each file stays ≤230 lines without changing player-visible behavior.
