+++ app/lib/features/game/flame/map_state/game_map_area_turn_feed_taps.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-
 import 'package:colonizethis_models/colonizethis_models.dart' as ct_models;
 
 import '../../../../config/routes.dart';
@@ -14,6 +13,7 @@ import 'map_location_resolver.dart';
 import 'game_map_area.dart';
 import 'game_map_area_state_base.dart';
 import 'game_map_area_turn_feed_labels.dart';
+import 'game_map_area_turn_feed_locate.dart';
 import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart';
 import 'package:colonizethis_logic/ai_api.dart';
 
@@ -22,7 +22,8 @@ mixin GameMapAreaTurnFeedTaps
     on
         ConsumerState<GameMapArea>,
         GameMapAreaStateBase,
-        GameMapAreaTurnFeedLabels {
+        GameMapAreaTurnFeedLabels,
+        GameMapAreaTurnFeedLocate {
   void navigateToDiplomacyDetail(String factionId) {
     final kind = factionKindForId(factionId);
     if (kind == null) {
@@ -35,7 +36,9 @@ mixin GameMapAreaTurnFeedTaps
           factionId,
           widget.game.worldState.turnState.turnNumber,
         );
-    ref.read(appEventBusProvider).emit(
+    ref
+        .read(appEventBusProvider)
+        .emit(
           ct_models.NavigateToRouteEvent(Routes.diplomacyDetail, {
             'game': widget.game,
             'humanPlayerId': mapPlayerId,
@@ -72,9 +75,9 @@ mixin GameMapAreaTurnFeedTaps
       return;
     }
     locateProvinceTile(province);
-    ref.read(appEventBusProvider).emit(
-          ct_models.OpenMapTileDetailEvent(tileKey: tileKey),
-        );
+    ref
+        .read(appEventBusProvider)
+        .emit(ct_models.OpenMapTileDetailEvent(tileKey: tileKey));
   }
 
   void Function()? provinceOverlayTapForProvince(String provinceId) {
@@ -96,9 +99,9 @@ mixin GameMapAreaTurnFeedTaps
     }
     return () {
       locateSeaZoneTile(seaZoneId);
-      ref.read(appEventBusProvider).emit(
-            ct_models.OpenMapTileDetailEvent(tileKey: tileKey),
-          );
+      ref
+          .read(appEventBusProvider)
+          .emit(ct_models.OpenMapTileDetailEvent(tileKey: tileKey));
     };
   }
 
@@ -111,7 +114,9 @@ mixin GameMapAreaTurnFeedTaps
     }
     return () {
       locateTileKey(targetTileKey);
-      ref.read(appEventBusProvider).emit(
+      ref
+          .read(appEventBusProvider)
+          .emit(
             ct_models.OpenCivilianUnitsPanelEvent(
               initialSelectedUnitId: unitId,
             ),
@@ -120,7 +125,9 @@ mixin GameMapAreaTurnFeedTaps
   }
 
   void Function()? overseasProfitCreditedTap() {
-    return () => ref.read(appEventBusProvider).emit(
+    return () => ref
+        .read(appEventBusProvider)
+        .emit(
           ct_models.NavigateToRouteEvent(Routes.trade, {
             'game': widget.game,
             'humanPlayerId': mapPlayerId,
@@ -130,7 +137,9 @@ mixin GameMapAreaTurnFeedTaps
   }
 
   void Function()? productionPanelTap() {
-    return () => ref.read(appEventBusProvider).emit(
+    return () => ref
+        .read(appEventBusProvider)
+        .emit(
           ct_models.NavigateToRouteEvent(Routes.production, {
             'game': widget.game,
             'humanPlayerId': mapPlayerId,
@@ -144,23 +153,25 @@ mixin GameMapAreaTurnFeedTaps
     switch (orderKind) {
       case ct_models.OrderKind.work:
       case ct_models.OrderKind.recruitWorker:
-        return () => ref.read(appEventBusProvider).emit(
-              const ct_models.OpenCivilianUnitsPanelEvent(),
-            );
+        return () => ref
+            .read(appEventBusProvider)
+            .emit(const ct_models.OpenCivilianUnitsPanelEvent());
       case ct_models.OrderKind.move:
       case ct_models.OrderKind.armyMove:
-        return () => ref.read(appEventBusProvider).emit(
-              const ct_models.OpenMilitaryUnitsPanelEvent(),
-            );
+        return () => ref
+            .read(appEventBusProvider)
+            .emit(const ct_models.OpenMilitaryUnitsPanelEvent());
       case ct_models.OrderKind.navalMove:
       case ct_models.OrderKind.navalMission:
-        return () => ref.read(appEventBusProvider).emit(
-              const ct_models.OpenNavalUnitsPanelEvent(),
-            );
+        return () => ref
+            .read(appEventBusProvider)
+            .emit(const ct_models.OpenNavalUnitsPanelEvent());
       case ct_models.OrderKind.buildUnit:
         return productionPanelTap();
       case ct_models.OrderKind.trade:
-        return () => ref.read(appEventBusProvider).emit(
+        return () => ref
+            .read(appEventBusProvider)
+            .emit(
               ct_models.NavigateToRouteEvent(Routes.trade, {
                 'game': widget.game,
                 'humanPlayerId': mapPlayerId,
@@ -173,7 +184,9 @@ mixin GameMapAreaTurnFeedTaps
         if (topology == null) {
           return null;
         }
-        return () => ref.read(appEventBusProvider).emit(
+        return () => ref
+            .read(appEventBusProvider)
+            .emit(
               ct_models.NavigateToRouteEvent(Routes.diplomacy, {
                 'game': widget.game,
                 'humanPlayerId': mapPlayerId,
