+++ app/lib/features/game/screens/counsel/counsel_screen_callbacks.dart
@@ -6,22 +6,14 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/industry_counsel_api.dart'
     show TradeCounselBookResult;
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:colonizethis_orders/colonizethis_orders.dart' show OrderEngine;
-import 'package:colonizethis_turn/colonizethis_turn.dart'
-    show projectOrderEffects;
 import 'package:flutter/material.dart';
 
-import '../../../../config/routes.dart';
-import 'counsel_development_apply.dart';
 import 'counsel_development_tab_body.dart';
-import 'counsel_industry_apply.dart';
 import 'counsel_industry_tab_body.dart';
-import 'counsel_military_apply.dart';
-import 'counsel_military_invade_confirm.dart';
 import 'counsel_military_tab_body.dart';
-import 'counsel_trade_apply.dart';
+import 'counsel_screen_callbacks_industry.dart';
+import 'counsel_screen_callbacks_military.dart';
 import 'counsel_trade_tab_body.dart';
-import 'military_counsel_l10n.dart';
 
 /// Tab callback objects wired from [CounselScreen] bodyBuilder (Refs #4352).
 final class CounselScreenTabCallbacks {
@@ -55,201 +47,48 @@ CounselScreenTabCallbacks buildCounselScreenTabCallbacks({
   required bool canEdit,
   required TradeCounselBookResult tradeCounsel,
 }) {
-  final industryCallbacks = CounselIndustryCallbacks(
-    onApplyProduceAllocation: canEdit
-        ? () {
-            final currentDesired = readProductionDesiredOutput();
-            final next = industryCounselDesiredOutputAfterProduceAgree(
-              game: displayGame,
-              playerId: humanPlayerId,
-              currentDesired: currentDesired,
-            );
-            replaceProductionDesiredOutput(next);
-          }
-        : null,
-    onAgreeTrain: canEdit
-        ? (tier) {
-            final orders = readCurrentOrders();
-            final next = industryCounselOrdersAfterTrainAgree(
-              currentOrders: orders,
-              playerId: humanPlayerId,
-              tier: tier,
-              game: displayGame,
-              topology: topology,
-            );
-            if (next == null) {
-              bus.emit(
-                ShowSnackBarEvent(
-                  message: l10n.industryCounsel_trainAgreeFailed,
-                ),
-              );
-              return;
-            }
-            replaceCurrentOrders(next);
-          }
-        : null,
-    onOpenDevelopment: canEdit
-        ? () {
-            bus.emit(
-              NavigateToRouteEvent(Routes.development, {
-                'game': displayGame,
-                'humanPlayerId': humanPlayerId,
-              }),
-            );
-          }
-        : null,
-  );
-  final tradeCallbacks = CounselTradeCallbacks(
-    onApplyBook: canEdit && tradeCounsel.book.isNotEmpty
-        ? () {
-            final orders = readCurrentOrders();
-            final next = tradeCounselOrdersAfterApplyBook(
-              currentOrders: orders,
-              playerId: humanPlayerId,
-              book: tradeCounsel.book,
-            );
-            replaceCurrentOrders(next);
-          }
-        : null,
-    onAgreeLine: canEdit
-        ? (order) {
-            final orders = readCurrentOrders();
-            final next = tradeCounselOrdersAfterAgree(
-              currentOrders: orders,
-              playerId: humanPlayerId,
-              order: order,
-            );
-            if (next == null) {
-              bus.emit(
-                ShowSnackBarEvent(message: l10n.tradeCounsel_applyFailed),
-              );
-              return;
-            }
-            replaceCurrentOrders(next);
-          }
-        : null,
-  );
-  final militaryCallbacks = CounselMilitaryCallbacks(
-    onAgreeTrain: canEdit
-        ? (recommendation) {
-            final unitType = recommendation.unitType;
-            final count = recommendation.count;
-            if (unitType == null || count == null || count <= 0) {
-              return;
-            }
-            final orders = readCurrentOrders();
-            final next = militaryCounselOrdersAfterTrainAgree(
-              game: displayGame,
-              playerId: humanPlayerId,
-              currentOrders: orders,
-              topology: topology,
-              unitType: unitType,
-              count: count,
-            );
-            if (next == null) {
-              bus.emit(
-                ShowSnackBarEvent(
-                  message: l10n.militaryCounsel_trainAgreeFailed,
-                ),
-              );
-              return;
-            }
-            replaceCurrentOrders(next);
-          }
-        : null,
-    onAgreeInvade: canEdit
-        ? (recommendation) async {
-            final destination =
-                militaryCounselInvadeDestinationForRecommendation(
-                  game: displayGame,
-                  playerId: humanPlayerId,
-                  currentOrders: readCurrentOrders(),
-                  topology: topology,
-                  recommendation: recommendation,
-                );
-            if (destination == null) {
-              bus.emit(
-                ShowSnackBarEvent(
-                  message: l10n.militaryCounsel_invadeAgreeFailed,
-                ),
-              );
-              return;
-            }
-            if (destination.requiresDeclareWarOnConfirm) {
-              final ownerLabel = militaryCounselOwnerLabel(
-                l10n,
-                displayGame,
-                recommendation,
-              );
-              final ok = await showMilitaryCounselDeclareWarConfirmDialog(
-                context,
-                l10n,
-                ownerLabel,
-                game: displayGame,
-                humanPlayerId: humanPlayerId,
-                targetFactionId: destination.ownerFactionId,
-              );
-              if (ok != true || !context.mounted) return;
-            }
-            final orders = readCurrentOrders();
-            final next = militaryCounselOrdersAfterInvadeAgree(
-              currentOrders: orders,
-              playerId: humanPlayerId,
-              armyId: recommendation.armyId!,
-              destination: destination,
-            );
-            final engine = OrderEngine(
-              initialOrders: next,
-              projector: projectOrderEffects,
-            );
-            final results = engine.validatePlayerOrdersWithContext(
-              displayGame,
-              topology,
-              humanPlayerId,
-            );
-            if (!results.every((r) => r.isAccepted)) {
-              bus.emit(
-                ShowSnackBarEvent(
-                  message: l10n.militaryCounsel_invadeAgreeFailed,
-                ),
-              );
-              return;
-            }
-            replaceCurrentOrders(next);
-          }
-        : null,
-  );
-  final developmentCallbacks = CounselDevelopmentCallbacks(
-    onAgreeBuildPort: canEdit
-        ? (recommendation) {
-            final orders = readCurrentOrders();
-            final workOrder = developmentCounselPortWorkOrderAfterAgree(
-              game: displayGame,
-              playerId: humanPlayerId,
-              currentOrders: orders,
-              topology: topology,
-              recommendation: recommendation,
-              tileMapByRegion: tileMapByRegion,
-            );
-            if (workOrder == null) {
-              bus.emit(
-                ShowSnackBarEvent(message: l10n.developmentCounsel_agreeFailed),
-              );
-              return;
-            }
-            bus.emit(
-              UpsertPendingCivilianWorkOrderRequestedEvent(
-                playerId: humanPlayerId,
-                workOrder: workOrder,
-              ),
-            );
-          }
-        : null,
-  );
   return CounselScreenTabCallbacks(
-    industry: industryCallbacks,
-    trade: tradeCallbacks,
-    military: militaryCallbacks,
-    development: developmentCallbacks,
+    industry: buildCounselIndustryCallbacks(
+      bus: bus,
+      readCurrentOrders: readCurrentOrders,
+      replaceCurrentOrders: replaceCurrentOrders,
+      readProductionDesiredOutput: readProductionDesiredOutput,
+      replaceProductionDesiredOutput: replaceProductionDesiredOutput,
+      displayGame: displayGame,
+      humanPlayerId: humanPlayerId,
+      topology: topology,
+      l10n: l10n,
+      canEdit: canEdit,
+    ),
+    trade: buildCounselTradeCallbacks(
+      bus: bus,
+      readCurrentOrders: readCurrentOrders,
+      replaceCurrentOrders: replaceCurrentOrders,
+      humanPlayerId: humanPlayerId,
+      l10n: l10n,
+      canEdit: canEdit,
+      tradeCounsel: tradeCounsel,
+    ),
+    military: buildCounselMilitaryCallbacks(
+      context: context,
+      bus: bus,
+      readCurrentOrders: readCurrentOrders,
+      replaceCurrentOrders: replaceCurrentOrders,
+      displayGame: displayGame,
+      humanPlayerId: humanPlayerId,
+      topology: topology,
+      l10n: l10n,
+      canEdit: canEdit,
+    ),
+    development: buildCounselDevelopmentCallbacks(
+      bus: bus,
+      readCurrentOrders: readCurrentOrders,
+      displayGame: displayGame,
+      humanPlayerId: humanPlayerId,
+      topology: topology,
+      tileMapByRegion: tileMapByRegion,
+      l10n: l10n,
+      canEdit: canEdit,
+    ),
   );
 }
