+++ app/lib/features/game/screens/trade/trade_screen_market_tab_build.dart
@@ -2,7 +2,6 @@
 // Split from `trade_screen_market_tab.dart` to keep each trade-screen part
 // under the repo file-size target (Refs #3878).
 
-
 import 'package:colonizethis_app_l10n/l10n/l10n.dart';
 import 'package:flutter/material.dart';
 
@@ -12,23 +11,17 @@ import 'package:colonizethis_economy/colonizethis_economy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../../../../providers/games_provider.dart';
-import '../../../../widgets/ct_nine_patch_button.dart';
 
 import 'trade_market_staging_context.dart';
 import 'trade_screen_contract_market.dart';
 import 'trade_screen_market_tab.dart';
+import 'trade_screen_market_tab_build_chrome.dart';
 import 'trade_screen_market_tab_build_sections.dart';
-import 'trade_screen_market_tab_cargo_header.dart';
 import 'trade_screen_market_tab_catalog.dart';
 import 'trade_screen_market_tab_order_handlers.dart';
 import 'trade_section_handlers.dart';
 
 extension MarketTabContentBuild on MarketTabContent {
-  /// Visual dim factor applied to the Market tab body when the screen
-  /// is in observe mode (`canMutateViaUi == false`). Matches the
-  /// editorial-monocle conventions for read-only surfaces.
-  static const double observeModeOpacity = 0.7;
-
   Widget buildMarketTabBody(
     BuildContext context, {
     required Orders orders,
@@ -36,7 +29,7 @@ extension MarketTabContentBuild on MarketTabContent {
     required Map<String, int> desiredOutputByRecipe,
     required int? Function() readProjectedTreasuryDelta,
     required Map<CommodityId, TradeCounselRecommendation>
-        tradeCounselHighlightsByCommodityId,
+    tradeCounselHighlightsByCommodityId,
     TradeOpenCounselCallback? onOpenTradeCounsel,
   }) {
     final ThemeData theme = Theme.of(context);
@@ -60,12 +53,12 @@ extension MarketTabContentBuild on MarketTabContent {
 
     final TradeSectionHandlers sectionHandlers = buildTradeSectionHandlers(
       readProjectedTreasuryDelta: readProjectedTreasuryDelta,
-      handleDirectionChanged: ({
-        required CommodityId commodityId,
-        required TradeOrderType? next,
-        required int? projectedTreasuryDelta,
-      }) =>
-          handleDirectionChanged(
+      handleDirectionChanged:
+          ({
+            required CommodityId commodityId,
+            required TradeOrderType? next,
+            required int? projectedTreasuryDelta,
+          }) => handleDirectionChanged(
             ordersNotifier: ordersNotifier,
             orders: orders,
             productionInputConsumption: productionInputConsumption,
@@ -73,12 +66,12 @@ extension MarketTabContentBuild on MarketTabContent {
             commodityId: commodityId,
             next: next,
           ),
-      handleQuantityDelta: ({
-        required CommodityId commodityId,
-        required int delta,
-        required int? projectedTreasuryDelta,
-      }) =>
-          handleQuantityDelta(
+      handleQuantityDelta:
+          ({
+            required CommodityId commodityId,
+            required int delta,
+            required int? projectedTreasuryDelta,
+          }) => handleQuantityDelta(
             ordersNotifier: ordersNotifier,
             orders: orders,
             productionInputConsumption: productionInputConsumption,
@@ -87,31 +80,31 @@ extension MarketTabContentBuild on MarketTabContent {
             delta: delta,
           ),
     );
-    final TradeMarketStagingContext staging = TradeMarketStagingContext.forMarketBuild(
-      game: game,
-      playerId: playerId,
-      orders: orders,
-      productionInputConsumption: productionInputConsumption,
-      handlers: sectionHandlers,
-      tradeCounselHighlightsByCommodityId: tradeCounselHighlightsByCommodityId,
-      onOpenTradeCounsel: onOpenTradeCounsel,
-    );
+    final TradeMarketStagingContext staging =
+        TradeMarketStagingContext.forMarketBuild(
+          game: game,
+          playerId: playerId,
+          orders: orders,
+          productionInputConsumption: productionInputConsumption,
+          handlers: sectionHandlers,
+          tradeCounselHighlightsByCommodityId:
+              tradeCounselHighlightsByCommodityId,
+          onOpenTradeCounsel: onOpenTradeCounsel,
+        );
 
     final int tradeCargoCapacity = cargoHoldsForHomeFleet(game, playerId);
     final int totalStagedBid = totalStagedBidQuantity(orders, playerId);
-    final int stagedDistinctBidCount =
-        stagedDistinctBidCommodityCount(orders, playerId);
+    final int stagedDistinctBidCount = stagedDistinctBidCommodityCount(
+      orders,
+      playerId,
+    );
     final int bidTypeCap = staging.bidTypeCap;
     final int remainingCargo = tradeCargoCapacity - totalStagedBid;
     final int clampedRemaining = remainingCargo < 0 ? 0 : remainingCargo;
-    final bool warningVisible =
-        clampedRemaining == 0 && totalStagedBid > 0;
+    final bool warningVisible = clampedRemaining == 0 && totalStagedBid > 0;
 
-    final ({
-      int budgetTotal,
-      int budgetRemaining,
-      bool warningVisible,
-    }) bidBudgetHeader = marketTabBidBudgetHeaderState(
+    final ({int budgetTotal, int budgetRemaining, bool warningVisible})
+    bidBudgetHeader = marketTabBidBudgetHeaderState(
       game: game,
       playerId: playerId,
       orders: orders,
@@ -138,7 +131,8 @@ extension MarketTabContentBuild on MarketTabContent {
     // row-major two-column grid per section with compact two-line rows.
     final Widget list = LayoutBuilder(
       builder: (BuildContext context, BoxConstraints constraints) {
-        final bool wideLayout = constraints.maxWidth >=
+        final bool wideLayout =
+            constraints.maxWidth >=
             TradeScreenMarketKeys.marketTwoColumnMinWidth;
         final List<Widget> sectionWidgets = buildMarketTabSectionWidgets(
           l10n: l10n,
@@ -162,40 +156,20 @@ extension MarketTabContentBuild on MarketTabContent {
       },
     );
 
-    final Widget header = RepaintBoundary(
-      key: TradeScreenMarketKeys.marketBidTypeCapGoldenKey,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          if (onOpenTradeCounsel != null)
-            Align(
-              alignment: Alignment.centerRight,
-              child: Padding(
-                padding: const EdgeInsets.fromLTRB(12, 8, 12, 0),
-                child: CtNinePatchButton(
-                  key: const ValueKey<String>('trade_market_counsel_button'),
-                  onPressed: () => onOpenTradeCounsel(),
-                  child: Text(l10n.tradeMarket_counsel),
-                ),
-              ),
-            ),
-          MarketTabHeaderStrip(
-            stagedDistinctBidCount: stagedDistinctBidCount,
-            bidTypeCap: bidTypeCap,
-            clampedRemaining: clampedRemaining,
-            cargoWarningVisible: warningVisible,
-            bidBudgetTotal: bidBudgetHeader.budgetTotal,
-            bidBudgetRemaining: bidBudgetHeader.budgetRemaining,
-            bidBudgetWarningVisible: bidBudgetHeader.warningVisible,
-            bidGoodsIndicatorStyle: bidGoodsIndicatorStyle,
-            bidTypeWarningStyle: bidTypeWarningStyle,
-            cargoIndicatorStyle: cargoIndicatorStyle,
-            cargoWarningStyle: cargoWarningStyle,
-            bidBudgetIndicatorStyle: bidBudgetIndicatorStyle,
-            bidBudgetWarningStyle: bidBudgetWarningStyle,
-          ),
-        ],
-      ),
+    final Widget header = buildMarketTabCounselAndHeader(
+      l10n: l10n,
+      stagedDistinctBidCount: stagedDistinctBidCount,
+      bidTypeCap: bidTypeCap,
+      clampedRemaining: clampedRemaining,
+      cargoWarningVisible: warningVisible,
+      bidBudgetHeader: bidBudgetHeader,
+      bidGoodsIndicatorStyle: bidGoodsIndicatorStyle,
+      bidTypeWarningStyle: bidTypeWarningStyle,
+      cargoIndicatorStyle: cargoIndicatorStyle,
+      cargoWarningStyle: cargoWarningStyle,
+      bidBudgetIndicatorStyle: bidBudgetIndicatorStyle,
+      bidBudgetWarningStyle: bidBudgetWarningStyle,
+      onOpenTradeCounsel: onOpenTradeCounsel,
     );
 
     final Widget body = Column(
@@ -216,19 +190,7 @@ extension MarketTabContentBuild on MarketTabContent {
     // read-only, but leave the bid-goods + cargo header live (they're
     // read-only telemetry that should still surface state).
     if (!canEdit) {
-      return Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          header,
-          Flexible(
-            child: Opacity(
-              opacity: observeModeOpacity,
-              child: IgnorePointer(child: list),
-            ),
-          ),
-        ],
-      );
+      return wrapMarketTabObserveList(header: header, list: list);
     }
     return body;
   }
@@ -243,8 +205,9 @@ extension MarketTabContentBuild on MarketTabContent {
     if (desiredOutputByRecipe.isEmpty) {
       return const <CommodityId, int>{};
     }
-    final List<AssignedRecipe> assignments =
-        assignedRecipesFromDesiredOutput(desiredOutputByRecipe);
+    final List<AssignedRecipe> assignments = assignedRecipesFromDesiredOutput(
+      desiredOutputByRecipe,
+    );
     if (assignments.isEmpty) {
       return const <CommodityId, int>{};
     }
