+++ app/lib/features/game/widgets/diplomacy/diplomacy_panel_row.dart
@@ -1,22 +1,16 @@
 // Per-faction diplomacy row layout for DiplomacyPanel.
 // SPEC/ui/diplomacy-panel.md § Per-faction row, § Responsive layout.
 
-import 'package:colonizethis_app_l10n/l10n/l10n.dart';
-import 'package:colonizethis_app_ui_chrome/config/editorial_monocle_palette.dart';
-import 'package:colonizethis_diplomacy/colonizethis_diplomacy.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../widgets/ct_gap.dart';
 import '../../../../widgets/ct_spacing.dart';
-import '../../../../widgets/relation_meter.dart';
-import 'diplomacy_panel_chrome_relation_badges.dart';
-import 'diplomacy_panel_chrome_section_header.dart';
 import 'diplomacy_panel_chrome_standing.dart';
 import 'diplomacy_panel_constants.dart';
 import 'diplomacy_panel_row_actions.dart';
+import 'diplomacy_panel_row_info.dart';
 import 'diplomacy_panel_rows.dart';
-import 'relative_power_line.dart';
 
 class DiplomacyRow extends StatelessWidget {
   const DiplomacyRow({
@@ -54,7 +48,7 @@ class DiplomacyRow extends StatelessWidget {
       key: _bodyKey,
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Expanded(child: _buildInfoColumn(context)),
+        Expanded(child: DiplomacyRowInfo(data: data)),
         Flexible(
           child: Align(
             alignment: Alignment.topRight,
@@ -72,174 +66,12 @@ class DiplomacyRow extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisSize: MainAxisSize.min,
       children: [
-        _buildInfoColumn(context),
+        DiplomacyRowInfo(data: data),
         if (hasActions) ...[
           CtGap.m,
           Align(alignment: Alignment.centerLeft, child: _buildActionButtons()),
         ],
       ],
-    );
-  }
-
-  Widget _buildInfoColumn(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        _buildHeaderRow(context),
-        ..._buildRelativePowerLine(context),
-        const SizedBox(height: 4),
-        _buildRelationRow(context),
-        ..._buildStandingChips(context),
-        ..._buildOptionalStatusLines(context),
-      ],
-    );
-  }
-
-  Widget _buildHeaderRow(BuildContext context) {
-    return Row(
-      children: [
-        Flexible(
-          child: Text(
-            data.displayName,
-            style: Theme.of(
-              context,
-            ).textTheme.titleSmall?.copyWith(fontWeight: FontWeight.w600),
-            overflow: TextOverflow.ellipsis,
-            maxLines: 1,
-          ),
-        ),
-        CtGap.wm,
-        DiplomacyFactionKindBadge(kind: data.kind),
-      ],
-    );
-  }
-
-  List<Widget> _buildRelativePowerLine(BuildContext context) {
-    final int? gpScore = data.powerScore;
-    final int? playerScore = data.playerPowerScore;
-    if (gpScore == null || playerScore == null) {
-      return const [];
-    }
-    final int pct = powerComparisonPercent(gpScore, playerScore);
-    return [const SizedBox(height: 4), RelativePowerLine(pct: pct)];
-  }
-
-  Widget _buildRelationRow(BuildContext context) {
-    final TextStyle bodySmall =
-        Theme.of(context).textTheme.bodySmall ?? const TextStyle(fontSize: 12);
-    final DiplomacyRelation? rel = data.relation;
-    if (rel == null) {
-      return Text('—', style: bodySmall);
-    }
-    final String relationStateLabel = relationScoreToDisplayLabel(rel.score);
-    final TextStyle mutedStyle = bodySmall.copyWith(
-      color: EditorialMonoclePalette.muted,
-    );
-    final TextStyle wordStyle = bodySmall.copyWith(
-      color: diplomacyRelationWordColor(rel.score),
-      fontStyle: FontStyle.italic,
-    );
-    final List<InlineSpan> spans = <InlineSpan>[];
-    if (relationStateLabel.isNotEmpty) {
-      spans.add(TextSpan(text: relationStateLabel, style: wordStyle));
-    }
-    final bool showAlliance = rel.formalAlliance;
-    return Wrap(
-      crossAxisAlignment: WrapCrossAlignment.center,
-      spacing: CtSpacing.s,
-      runSpacing: CtSpacing.xs,
-      children: [
-        DiplomacyRelationStateBadge(atWar: rel.atWar),
-        if (showAlliance) const DiplomacyAllianceBadge(),
-        RelationMeter(score: rel.score),
-        if (spans.isNotEmpty)
-          Text.rich(
-            TextSpan(style: mutedStyle, children: spans),
-            overflow: TextOverflow.ellipsis,
-            maxLines: 2,
-          ),
-      ],
-    );
-  }
-
-  List<Widget> _buildStandingChips(BuildContext context) {
-    if (data.standingChips.isEmpty) {
-      return const <Widget>[];
-    }
-    return <Widget>[
-      const SizedBox(height: 4),
-      DiplomacyStandingChipCluster(chips: data.standingChips),
-    ];
-  }
-
-  TextStyle _economicLineStyle(BuildContext context) {
-    final TextStyle base =
-        Theme.of(context).textTheme.bodySmall ?? const TextStyle(fontSize: 12);
-    return base.copyWith(
-      color: EditorialMonoclePalette.accentDim,
-      fontFamily: 'monospace',
-      fontFamilyFallback: const ['Courier'],
-      fontStyle: FontStyle.normal,
-    );
-  }
-
-  List<Widget> _buildOptionalStatusLines(BuildContext context) {
-    final l10n = appL10n(context);
-    final TextStyle style = _economicLineStyle(context);
-    final lines = <Widget>[];
-    if (data.activeSubsidyPercent != null) {
-      lines.addAll([
-        const SizedBox(height: 4),
-        _subsidyEconomicLine(
-          compact: l10n.diplomacy_panel_outgoingSubsidy(
-            data.activeSubsidyPercent!,
-            data.displayName,
-          ),
-          percent: data.activeSubsidyPercent!,
-          lineKey: const Key('diplomacyOutgoingSubsidyLine'),
-          style: style,
-        ),
-      ]);
-    }
-    if (data.pendingGrantAmount != null) {
-      lines.addAll([
-        const SizedBox(height: 4),
-        Text(
-          l10n.diplomacy_panel_pendingGrant(data.pendingGrantAmount!),
-          style: style,
-        ),
-      ]);
-    }
-    if (data.pendingSubsidyPercent != null) {
-      lines.addAll([
-        const SizedBox(height: 4),
-        _subsidyEconomicLine(
-          compact: l10n.diplomacy_panel_pendingSubsidy(
-            data.pendingSubsidyPercent!,
-          ),
-          percent: data.pendingSubsidyPercent!,
-          lineKey: const Key('diplomacyPendingSubsidyLine'),
-          style: style,
-        ),
-      ]);
-    }
-    return lines;
-  }
-
-  Widget _subsidyEconomicLine({
-    required String compact,
-    required int percent,
-    required Key lineKey,
-    required TextStyle style,
-  }) {
-    final tooltip = subsidyFillPriceConsequenceTooltip(
-      targetDisplayName: data.displayName,
-      percent: percent,
-    );
-    return Tooltip(
-      message: tooltip,
-      child: Text(compact, key: lineKey, style: style, semanticsLabel: tooltip),
     );
   }
