+++ app/lib/features/game/widgets/units/naval/naval_units_panel_list.dart
@@ -11,12 +11,14 @@ import '../shared/region_section_header.dart';
 import 'naval_units_panel.dart';
 import 'naval_units_panel_state_base.dart';
 import 'naval_units_panel_support_combine.dart';
+import 'naval_units_panel_support_combine_home.dart';
 import 'naval_units_panel_support_dialogs.dart';
 
 mixin NavalUnitsPanelList
     on
         BaseUnitsPanelState<NavalUnitsPanel>,
         NavalUnitsPanelStateBase,
+        NavalUnitsPanelCombineHome,
         NavalUnitsPanelCombine,
         NavalUnitsPanelDialogs {
   List<Widget> navalListChildren({
