+++ app/lib/features/game/widgets/units/naval/naval_units_panel_support_combine.dart
@@ -1,24 +1,18 @@
-import 'package:colonizethis_app/core/utils/prefixed_id.dart';
-import 'package:colonizethis_world/colonizethis_world.dart'
-    show GamePlayerLookup, homeFleetIdFor;
+import 'package:colonizethis_world/colonizethis_world.dart' show homeFleetIdFor;
 
 import 'package:colonizethis_models/colonizethis_models.dart';
-import 'package:flutter/material.dart';
 
 import '../../panels/tree_builders/naval_tree_builder.dart';
-import '../../unit_orders/transfer_to_home_fleet_dialog.dart';
 import '../shared/base_units_panel.dart';
 import 'naval_units_panel.dart';
 import 'naval_units_panel_state_base.dart';
+import 'naval_units_panel_support_combine_home.dart';
 
 mixin NavalUnitsPanelCombine
-    on BaseUnitsPanelState<NavalUnitsPanel>, NavalUnitsPanelStateBase {
-  /// Canonical fleet id for combine/split selection (Home Fleet uses [homeFleetIdFor]).
-  String selectionFleetId(FleetRow row) {
-    if (row.isHomeFleet) return homeFleetIdFor(widget.humanPlayerId);
-    return row.fleetId;
-  }
-
+    on
+        BaseUnitsPanelState<NavalUnitsPanel>,
+        NavalUnitsPanelStateBase,
+        NavalUnitsPanelCombineHome {
   bool canCombineSelection(List<FleetRow> flat) {
     final rowsById = <String, FleetRow>{
       for (final r in flat) selectionFleetId(r): r,
@@ -51,25 +45,6 @@ mixin NavalUnitsPanelCombine
       if (selected.contains(id)) return id;
     }
     throw StateError('combine target: empty selection');
-  }
-
-  Fleet? fleetForRow(FleetRow row) {
-    final id = selectionFleetId(row);
-    final found = widget.game.fleetById(id);
-    if (found != null) return found;
-    if (row.isHomeFleet) {
-      final portId = row.inPortAtProvinceId;
-      if (portId == null) return null;
-      return Fleet(
-        id: id,
-        ownerId: widget.humanPlayerId,
-        regionId: row.regionId,
-        inPortAtProvinceId: portId,
-        ships: const [],
-        mission: FleetMission.none,
-      );
-    }
-    return null;
   }
 
   void toggleFleetSelection(FleetRow row) {
@@ -165,91 +140,5 @@ mixin NavalUnitsPanelCombine
     }
     if (home == null || source == null) return null;
     return (home: home, source: source);
-  }
-
-  String? humanCapitalProvinceId() =>
-      widget.game.playerById(widget.humanPlayerId)?.capitalProvinceId;
-
-  bool provinceMatchesCapital(String provinceId, String capitalProvinceId) {
-    if (provinceId == capitalProvinceId) return true;
-    final capRegionId = ProvinceId.regionIdFrom(capitalProvinceId);
-    final capLocalId = ProvinceId.localIdFrom(capitalProvinceId);
-    return provinceId == capLocalId || provinceId == '$capRegionId|$capLocalId';
-  }
-
-  bool seaZoneAdjacentToCapital({
-    required String sourceSeaZoneId,
-    required String sourceRegionId,
-    required String capitalProvinceId,
-  }) {
-    final capRegionId = ProvinceId.regionIdFrom(capitalProvinceId);
-    final capLocalId = ProvinceId.localIdFrom(capitalProvinceId);
-    final sourceSeaLocal = prefixedIdLocalSegment(sourceSeaZoneId);
-    final sourceSeaPrefixed = prefixedIdHasDelimiter(sourceSeaZoneId)
-        ? sourceSeaZoneId
-        : '$sourceRegionId|$sourceSeaZoneId';
-    final sourceSeaCandidates = <String>{
-      sourceSeaZoneId,
-      sourceSeaLocal,
-      sourceSeaPrefixed,
-    };
-    final capitalCandidates = <String>{
-      capitalProvinceId,
-      capLocalId,
-      '$capRegionId|$capLocalId',
-    };
-    for (final edge in widget.topology.edges) {
-      final a = edge.id1;
-      final b = edge.id2;
-      final aIsSea = sourceSeaCandidates.contains(a);
-      final bIsSea = sourceSeaCandidates.contains(b);
-      final aIsCap = capitalCandidates.contains(a);
-      final bIsCap = capitalCandidates.contains(b);
-      if ((aIsSea && bIsCap) || (bIsSea && aIsCap)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  bool isEligibleHomeTransferSource(FleetRow sourceRow) {
-    final sourceFleet = fleetForRow(sourceRow);
-    final capitalProvinceId = humanCapitalProvinceId();
-    if (sourceFleet == null || capitalProvinceId == null) return false;
-    if (sourceFleet.ownerId != widget.humanPlayerId) return false;
-    if (!sourceFleet.isAtSea) {
-      final inPortId = sourceFleet.inPortAtProvinceId;
-      if (inPortId == null) return false;
-      return provinceMatchesCapital(inPortId, capitalProvinceId);
-    }
-    final seaZoneId = sourceFleet.seaZoneId;
-    if (seaZoneId == null || seaZoneId.isEmpty) return false;
-    return seaZoneAdjacentToCapital(
-      sourceSeaZoneId: seaZoneId,
-      sourceRegionId: sourceFleet.regionId,
-      capitalProvinceId: capitalProvinceId,
-    );
-  }
-
-  void openTransferToHomeDialog({
-    required FleetRow homeRow,
-    required FleetRow sourceRow,
-  }) {
-    final homeFleet = fleetForRow(homeRow);
-    final sourceFleet = fleetForRow(sourceRow);
-    if (homeFleet == null || sourceFleet == null) return;
-    showDialog<void>(
-      context: context,
-      builder: (_) => TransferToHomeFleetDialog(
-        sourceFleet: sourceFleet,
-        homeFleet: homeFleet,
-        game: widget.game,
-        humanPlayerId: widget.humanPlayerId,
-        bus: widget.bus,
-        overseasCargoUsed: widget.overseasCargoUsed,
-        isCargoUsedReliable: widget.isCargoUsedReliable,
-        cargoNotDefined: widget.cargoNotDefined,
-      ),
-    );
   }
 }
